First Part of POC is now complete.
We are able to populate the database with questions (MC, Short Answers,
Numerical Answers). Please note, we focused on the Mobile Application
for the POC, thus not all desktop HTML are avaliable. Next Milestone
will be to let students answer questions. This will be completed before
our POC this Tuesday.