New Website Layout
This committ introduces the new website layout of InPress. We have used
the Twitter Bootstrap’s API in order to develop this new website
layout. This is fully compitable with all mobile devices (though
haven’t fully tested so far), and we will eventually (within the next
few days) tie this new website layout with the django web framework.